+++ digits/changelog.md
@@ -2,6 +2,8 @@
 *Non-trivial pull requests should include an entry below. Entries must be suitable for inclusion in public-facing materials such as release notes and blog posts. Keep them short, sweet, and in the past tense. New entries go on top. When merging to deploy, add the version number and date.*
 
 ## Unreleased
+
+* (EF) Use AuthRequestQueue for all API request to ensure we always have a valid guest auth token
 * (EF) Fixed critical issue where Digits sessions are lost when using Proguard.
 
 ## v1.6.1